changed names to .hpp files; edited unit conversions on vmax / amax

+++ src/miscDevices/LineSensorService.hpp
@@ -1,5 +1,7 @@
 #pragma once
 
+#include "LineSensor.hpp"
+
 namespace miscDevices
 {
     class LineSensorService
@@ -7,7 +9,7 @@ namespace miscDevices
         public:
             /// @brief determines the actual rotation of the vahicle based on the line position
             /// @param lineSensor current instance of the lineSensor
-            /// @return rotation in degree
-            static float getVehicleRotation(LineSensor *lineSensor);
+            /// @return rotation in 10 * degree
+            static int32_t getVehicleRotation(LineSensor *lineSensor);
     };
 }
